+++ objectbox/lib/src/native/store.dart
@@ -432,13 +432,23 @@ class Store {
   /// Spawns an isolate, runs [callback] in that isolate passing it [param] with
   /// its own Store and returns the result of callback.
   ///
+  /// ```dart
+  /// String? readNameAndRemove(Store store, int objectId) {
+  ///   var box = store.box<User>();
+  ///   final nameOrNull = box.get(objectId)?.name;
+  ///   box.remove(objectId);
+  ///   return nameOrNull;
+  /// }
+  /// await store.runAsync(readNameAndRemove, objectId);
+  /// ```
+  ///
   /// Instances of [callback] must be top-level functions or static methods
   /// of classes, not closures or instance methods of objects.
   ///
   /// Note: this requires Dart 2.15.0 or newer
   /// (shipped with Flutter 2.8.0 or newer).
-  Future<R> runIsolated<P, R>(
-      TxMode mode, FutureOr<R> Function(Store, P) callback, P param) async {
+  Future<R> runAsync<P, R>(
+      FutureOr<R> Function(Store, P) callback, P param) async {
     final resultPort = ReceivePort();
     final exitPort = ReceivePort();
     final errorPort = ReceivePort();
@@ -482,6 +492,21 @@ class Store {
     isolate.kill();
     return isolateResult.future;
   }
+
+  /// Deprecated. Use [runAsync] instead. Will be removed in a future release.
+  ///
+  /// Spawns an isolate, runs [callback] in that isolate passing it [param] with
+  /// its own Store and returns the result of callback.
+  ///
+  /// Instances of [callback] must be top-level functions or static methods
+  /// of classes, not closures or instance methods of objects.
+  ///
+  /// Note: this requires Dart 2.15.0 or newer
+  /// (shipped with Flutter 2.8.0 or newer).
+  @Deprecated('Use `runAsync` instead. Will be removed in a future release.')
+  Future<R> runIsolated<P, R>(TxMode mode,
+          FutureOr<R> Function(Store, P) callback, P param) async =>
+      runAsync(callback, param);
 
   /// Internal only - bypasses the main checks for async functions, you may
   /// only pass synchronous callbacks!